Simplifying ViewModel Declarations and Enhancing API and Application Efficiency with Filtering
In FavoritedCharactersView and RMCharacterView, the ViewModel declarations for the lists were moved to their respective page ViewModels.

Previously in RMCharactersView, there were two separate lists - the raw list and the filtered list - and their display in the view was controlled using if-else statements. This has been reduced to a single list. Additionally, the need for two separate API queries has been eliminated.

When filtering in RMCharactersView, a delay based on the inputs was added to the method triggering the API query. While waiting for the delay, a ProgressView is displayed instead of the list.