Updates in CharacterRowViewModel and CharacterServic
In CharacterRowViewModel, unnecessary database fetches were eliminated when checking if the character is favorite, now only the existence of the relevant character is checked.

In the CharacterService class, direct usage of the Filter object is adopted instead of 4 filter parameters, improving code readability and maintainability.

Refactored code in various sections for better readability and maintainability